cartpole/requirements.txt
appdirs==1.4.4
asttokens==2.4.1
av==12.0.0
backcall==0.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
executing==2.0.1
fonttools==4.50.0
gitdb==4.0.11
GitPython==3.1.43
gym==0.26.2
gym-notices==0.0.8
idna==3.6
imageio==2.23.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipdb==0.13.11
ipython==8.12.3
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
MinAtar==1.0.12
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.9.0.80
packaging==24.0
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pygame==2.1.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==1.44.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
threadpoolctl==3.4.0
tomli==2.0.1
torch==1.13.1
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.5
wcwidth==0.2.13
zipp==3.18.1
